+++ ncsa.hdf/src/ncsa/hdf/object/nc2/NC2Datatype.java
@@ -63,7 +63,7 @@ public class NC2Datatype extends Datatype {
     /*
      * (non-Javadoc)
      * 
-     * @see ncsa.hdf.object.DataFormat#hasAttribute()
+     * @see hdf.object.DataFormat#hasAttribute()
      */
     public boolean hasAttribute() {
         return false;
@@ -135,7 +135,7 @@ public class NC2Datatype extends Datatype {
      *            the netcdf native datatype.
      */
     @Override
-    public void fromNative(int tid) {
+    public void fromNative(long tid) {
         if (nativeType == null) {
             return;
         }
@@ -190,7 +190,7 @@ public class NC2Datatype extends Datatype {
 
     // implementing Datatype
     @Override
-    public int toNative() {
+    public long toNative() {
         if (datatypeClass == CLASS_INTEGER) {
             if (datatypeSize == 1) {
                 nativeType = DataType.BYTE;
@@ -223,10 +223,10 @@ public class NC2Datatype extends Datatype {
     /*
      * (non-Javadoc)
      * 
-     * @see ncsa.hdf.object.Datatype#close(int)
+     * @see hdf.object.Datatype#close(int)
      */
     @Override
-    public void close(int id) {
+    public void close(long id) {
         ;
     }
